% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabhelpers-package.R
\docType{package}
\name{tabhelpers-package}
\alias{tabhelpers}
\alias{tabhelpers-package}
\alias{class-package}
\title{tabhelpers: Helpers for displaying tables in command line}
\description{
Collection of functions to print the data frames in command line
}
\keyword{package}
