% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tabhelpers_options}
\alias{tabhelpers_options}
\title{Manage "tab_default" options}
\usage{
tabhelpers_options(o = NULL, value = NULL)
}
\arguments{
\item{o}{A character string specifying the option to retrieve or modify.}

\item{value}{The new value to assign to the specified option. If NULL, the function returns the current value.}
}
\value{
If no arguments are provided, it returns the entire "tab_default" list. If \code{o} is provided without \code{value}, it returns the current value of that option. If both \code{o} and \code{value} are provided, it updates the option.
}
\description{
This function allows retrieving or modifying the "tab_default" options.
If called without arguments, it returns the current list of options.
If \code{o} and \code{value} are provided, it updates the specified option.
}
\examples{
tabhelpers_options()  # Get all options
tabhelpers_options("vb_top")  # Get specific option
tabhelpers_options("border_char", "+")  # Modify an option

}
